requirements.txt
openai
python-magic
pdfminer.six
